+++ Cargo.toml
@@ -10,8 +10,6 @@ members = [
 	"runtime",
 	"runtime/common",
 	"tss",
-	"workers/task-executor",
-	"workers/time-worker",
 ]
 resolver = "2"
