+++ chronicle/Cargo.toml
@@ -4,15 +4,34 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = "1.0.75"
 async-channel = "1.9.0"
+async-trait = "0.1.73"
+bincode = "1.3.3"
 futures = "0.3.28"
 log = "0.4.20"
+rosetta-client = { git = "https://github.com/analog-labs/chain-connectors" }
 sc-client-api = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
 sc-network = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
 sc-transaction-pool-api = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
+serde = "1.0.188"
+serde_json = "1.0.105"
 sp-api = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
 sp-keystore = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
 sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-task-executor = { path = "../workers/task-executor" }
 time-primitives = { path = "../primitives" }
-time-worker = { path = "../workers/time-worker" }
+timegraph-client = "0.1.7"
+tokio = "1.32.0"
+tss = { path = "../tss" }
+
+[dev-dependencies]
+env_logger = "0.10.0"
+sc-consensus = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
+sc-network-test = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
+schnorr-evm = { git = "https://github.com/Analog-labs/frost-evm.git", version = "0.1.0" }
+surf = { version = "2.3.2", default-features = false, features = [ "h1-client-rustls" ] }
+sp-consensus = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
+sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0", default-features = false}
+sc-block-builder = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
+substrate-test-runtime-client = {  git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
+lazy_static = "1.4.0"
