--- workers/task-executor/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "task-executor"
-edition = "2021"
-version = "0.1.0"
-
-[dependencies]
-anyhow = "1.0.75"
-async-trait = "0.1.73"
-bincode = "1.3.3"
-futures = "0.3.28"
-log = "0.4.20"
-rosetta-client = { git = "https://github.com/analog-labs/chain-connectors" }
-sc-client-api = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0" }
-sc-transaction-pool-api = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-serde_json = "1.0.105"
-sp-api = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0" }
-sp-keystore = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v1.0.0" }
-time-primitives = { path = "../../primitives" }
-timegraph-client = "0.1.7"
-tokio = { version = "1.32.0", features = [ "macros" ] }
-
-[dev-dependencies]
-env_logger = "0.10.0"
-sc-network-test = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-surf = { version = "2.3.2", default-features = false, features = [ "h1-client-rustls" ] }
-sp-consensus = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0", default-features = false}
-sc-block-builder = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-substrate-test-runtime-client = {  git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-time-worker = {path = "../time-worker"}
-lazy_static = "1.4.0"
