--- workers/time-worker/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "time-worker"
-edition = "2021"
-version = "0.1.0"
-
-[dependencies]
-anyhow = "1.0.75"
-async-channel = "1.9.0"
-bincode = "1.3.3"
-futures = "0.3.28"
-log = "0.4"
-sc-client-api = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sc-network = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sc-transaction-pool-api = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-serde = { version = "1.0", features = [ "derive" ] }
-sp-api = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sp-keystore = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-time-primitives = { path = "../../primitives" }
-tokio = { version = "1.32.0", features = [ "time" ] }
-tss = { package = "tesseract-tss", path = "../../tss" }
-
-[dev-dependencies]
-async-trait = "0.1.73"
-env_logger = "0.10.0"
-frost-evm = { git = "https://github.com/Analog-labs/frost-evm.git", version = "0.1.0" }
-sc-consensus = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sc-network-test = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0", default-features = false}
-sp-consensus = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-substrate-test-runtime-client = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sc-transaction-pool = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0" }
-sp-std = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v1.0.0", default-features = false }
